WIP settings and scenarios/timelines
