+++ HumanSpaceProgram/Assets/HSP.Settings/SettingsFile.cs
@@ -6,12 +6,15 @@ namespace HSP.Settings
     /// <summary>
     /// The object representation of a settings file.
     /// </summary>
-    internal class SettingsFile
+    public class SettingsFile
     {
+        /// <summary>
+        /// The settings page instances in this file, with their current values.
+        /// </summary>
         public List<ISettingsPage> Pages;
 
         [MapsInheritingFrom( typeof( SettingsFile ) )]
-        public static SerializationMapping SettingsPage_GraphicsMapping()
+        public static SerializationMapping SettingsFileMapping()
         {
             return new MemberwiseSerializationMapping<SettingsFile>()
                 .WithMember( "pages", o => o.Pages );
