+++ HumanSpaceProgram/Assets/HSP._DevUtils/VanillaPlanetarySystemFactory.cs
@@ -181,7 +181,7 @@ namespace HSP._DevUtils
             return cb;
         }
 
-        [HSPEventListener( HSPEvent_BEFORE_TIMELINE_NEW.ID, CREATE_CELESTIAL_BODIES )]
+        //[HSPEventListener( HSPEvent_BEFORE_TIMELINE_NEW.ID, CREATE_CELESTIAL_BODIES )]
         //[HSPEventListener( HSPEvent_BEFORE_TIMELINE_LOAD.ID, CREATE_CELESTIAL_BODIES )]
         public static void CreateDefaultPlanetarySystem()
         {
