+++ HumanSpaceProgram/HSP._DevUtils.csproj
@@ -63,8 +63,8 @@
     <Compile Include="Assets\HSP._DevUtils\AIRFRotationFixer.cs" />
     <Compile Include="Assets\HSP._DevUtils\AsyncTester.cs" />
     <Compile Include="Assets\HSP._DevUtils\DevUtilsGameplayManager.cs" />
-    <Compile Include="Assets\HSP._DevUtils\ConstantForceApplier.cs" />
     <Compile Include="Assets\HSP._DevUtils\DevDefaultScenarioCreator.cs" />
+    <Compile Include="Assets\HSP._DevUtils\ConstantForceApplier.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Assets\HSP._DevUtils\WireframeSimpleQuad.shader" />
